feature/anthropic-batch-best-practices: address PR comments

+++ src/lambda-functions/migrations/0043_batch_executions.sql
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS batch_executions (
     status              TEXT NOT NULL DEFAULT 'submitted',
     submitted_at        TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     completed_at        TIMESTAMPTZ,
-    duration_ms         INTEGER,
+    duration_ms         BIGINT,
 
     request_count       INTEGER NOT NULL,
     succeeded_count     INTEGER,
